Feat/opcode classes (#492)
* Wip (does not work)

* Wip (still doesn't work)

* Wip (may seem like works but doesn't)

* Maybe works?

* Some fixes

* Fix

* Fix opcode shift somewhere

* chore: delete file leftover from merge

* vm tests pass

* chore: rename InstructionExecutorVariantName to ExecutorName

* Fix shift vm tests

* Remove unnecessary macro derives in instructions

* Disabled operation contains the opcode as a string now, not as a useless usize

* Program tests now have opcodes with default offset instead of just casting to usize

* Fix some compiler tests (not all), add some offset to compiler options (don't propagate them to the vm yet)

* Fix compiler tests, also make Core initialized by default in vm

* fix test in a hacky way

* chore: add comment

---------

+++ vm/src/arch/execution.rs
@@ -114,7 +114,7 @@ impl<F: Field> InstructionCols<F> {
             ..
         } = instruction;
         Self {
-            opcode: F::from_canonical_usize(*opcode as usize),
+            opcode: F::from_canonical_usize(*opcode),
             operands: [op_a, op_b, op_c, d, e, op_f, op_g].map(|&f| f),
         }
     }
